+++ demisto_sdk/commands/common/content/objects/pack_objects/pack_metadata/pack_metadata.py
@@ -6,16 +6,12 @@ from typing import Dict, List, Union
 from packaging.version import Version, parse
 from wcmatch.pathlib import Path
 
-from demisto_sdk.commands.common.constants import (PACKS_PACK_META_FILE_NAME,
-                                                   XSOAR_AUTHOR, XSOAR_SUPPORT,
-                                                   XSOAR_SUPPORT_URL,
-                                                   ContentItems)
-from demisto_sdk.commands.common.content.objects.abstract_objects import \
-    JSONObject
+from demisto_sdk.commands.common.constants import (PACKS_PACK_META_FILE_NAME, XSOAR_AUTHOR, XSOAR_SUPPORT,
+                                                   XSOAR_SUPPORT_URL, ContentItems)
+from demisto_sdk.commands.common.content.objects.abstract_objects import JSONObject
 from demisto_sdk.commands.common.handlers import JSON_Handler
 from demisto_sdk.commands.common.tools import get_core_pack_list
-from demisto_sdk.commands.find_dependencies.find_dependencies import \
-    PackDependencies
+from demisto_sdk.commands.find_dependencies.find_dependencies import PackDependencies
 
 json = JSON_Handler()