+++ demisto_sdk/commands/common/content/tests/objects/pack_objects/trigger/trigger_test.py
@@ -1,6 +1,5 @@
 from demisto_sdk.commands.common.content.objects.pack_objects import Trigger
-from demisto_sdk.commands.common.content.objects_factory import \
-    path_to_pack_object
+from demisto_sdk.commands.common.content.objects_factory import path_to_pack_object
 
 
 def get_trigger(pack, name):