+++ demisto_sdk/commands/content_graph/common.py
@@ -76,6 +76,7 @@ class ContentType(str, enum.Enum):
     XSIAM_DASHBOARD = "XSIAMDashboard"
     XSIAM_REPORT = "XSIAMReport"
     WIZARD = "Wizard"
+    XDRC_TEMPLATE = 'XDRCTemplate'
 
     @property
     def labels(self) -> List[str]: