+++ demisto_sdk/commands/content_graph/interface/neo4j/queries/constraints.py
@@ -1,9 +1,7 @@
 from neo4j import Transaction
 
-from demisto_sdk.commands.content_graph.common import (ContentType,
-                                                       RelationshipType)
-from demisto_sdk.commands.content_graph.interface.neo4j.queries.common import \
-    run_query
+from demisto_sdk.commands.content_graph.common import ContentType, RelationshipType
+from demisto_sdk.commands.content_graph.interface.neo4j.queries.common import run_query
 
 NODE_PROPERTY_UNIQUENESS_TEMPLATE = (
     "CREATE CONSTRAINT IF NOT EXISTS FOR (n:{label}) REQUIRE n.{prop} IS UNIQUE"