+++ demisto_sdk/commands/format/tests/test_formatting_readme_test.py
@@ -16,8 +16,7 @@ def test_format_wit_update_docker_flag(mocker):
     (but has a node), that the run ends without any exception.
     """
     from demisto_sdk.commands.common.git_util import GitUtil
-    from demisto_sdk.commands.common.hook_validations.readme import \
-        ReadMeValidator
+    from demisto_sdk.commands.common.hook_validations.readme import ReadMeValidator
     from demisto_sdk.commands.format.format_module import format_manager
     from demisto_sdk.commands.validate.validate_manager import ValidateManager
     mocker.patch.object(ReadMeValidator, 'are_modules_installed_for_verify', return_value=False)