+++ demisto_sdk/commands/integration_diff/integration_diff_detector.py
@@ -2,8 +2,7 @@ import os
 
 import click
 
-from demisto_sdk.commands.common.constants import (ARGUMENT_FIELDS_TO_CHECK,
-                                                   INTEGRATION_ARGUMENT_TYPES,
+from demisto_sdk.commands.common.constants import (ARGUMENT_FIELDS_TO_CHECK, INTEGRATION_ARGUMENT_TYPES,
                                                    PARAM_FIELDS_TO_CHECK)
 from demisto_sdk.commands.common.tools import get_yaml