+++ demisto_sdk/commands/split/jsonsplitter.py
@@ -2,12 +2,9 @@ import os
 
 import click
 
-from demisto_sdk.commands.common.constants import (DASHBOARDS_DIR,
-                                                   GENERIC_MODULES_DIR,
-                                                   PACKS_DIR)
+from demisto_sdk.commands.common.constants import DASHBOARDS_DIR, GENERIC_MODULES_DIR, PACKS_DIR
 from demisto_sdk.commands.common.handlers import JSON_Handler
-from demisto_sdk.commands.common.tools import (get_pack_name,
-                                               is_external_repository)
+from demisto_sdk.commands.common.tools import get_pack_name, is_external_repository
 
 json = JSON_Handler()