+++ demisto_sdk/commands/split/ymlsplitter.py
@@ -8,20 +8,14 @@ from io import open
 from pathlib import Path
 from typing import Optional
 
-from ruamel.yaml.scalarstring import (PlainScalarString,
-                                      SingleQuotedScalarString)
+from ruamel.yaml.scalarstring import PlainScalarString, SingleQuotedScalarString
 
 from demisto_sdk.commands.common.configuration import Configuration
-from demisto_sdk.commands.common.constants import (TYPE_PWSH, TYPE_PYTHON,
-                                                   TYPE_TO_EXTENSION)
+from demisto_sdk.commands.common.constants import TYPE_PWSH, TYPE_PYTHON, TYPE_TO_EXTENSION
 from demisto_sdk.commands.common.handlers import YAML_Handler
-from demisto_sdk.commands.common.tools import (LOG_COLORS,
-                                               get_all_docker_images,
-                                               get_pipenv_dir,
-                                               get_python_version, pascal_case,
-                                               print_color, print_error)
-from demisto_sdk.commands.unify.integration_script_unifier import \
-    IntegrationScriptUnifier
+from demisto_sdk.commands.common.tools import (LOG_COLORS, get_all_docker_images, get_pipenv_dir, get_python_version,
+                                               pascal_case, print_color, print_error)
+from demisto_sdk.commands.unify.integration_script_unifier import IntegrationScriptUnifier
 
 yaml = YAML_Handler()