+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 ## Unreleased
- * Fixed an issue where **update-release-notes** did not generate release notes for **XDRC Templates**.
- * Fixed an issue where **upload** failed without explaining the reason.
+* Added a validation that enforces predefined categories on MP Packs & integration yml files, the validation also ensures that each pack has only one category.
+* Fixed an issue where **update-release-notes** did not generate release notes for **XDRC Templates**.
+* Fixed an issue where **upload** failed without explaining the reason.
+* Improved implementation of the docker_helper module.
+* Fixed an issue where **validate** did not check changed pack_metadata.json files when running using git.
+* Added support for **xdrctemplate** to content graph.
+* Added a validation to validate types of attributes in the schema file of modeling rule.
+* Added a validation to validate that the dataset name is the same in the xif and schema files of modeling rule.
+* Added a validation to validate that a correlation rule file does not start with a hyphen.
 
 ## 1.7.8
 * Added the capability to run the MDX server in a docker container for environments without node.